old_tool/requirements_ubuntu18.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
astunparse==1.6.3
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
future==0.18.2
gast==0.2.2
gist==0.5
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
imutils==0.5.3
ipaddr==2.2.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.2.0
lockfile==0.12.2
Markdown==3.2.1
matplotlib==3.2.1
msgpack==0.6.2
numpy==1.18.4
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
opencv-python==3.4.2.16
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
sklearn==0.0
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.0
tqdm==4.46.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
